etl/requirements.txt
psycopg2-binary
elasticsearch==8.15.1
python-dotenv==1.0.1
pydantic==2.9.2
backoff==2.2.1